Add a `$` function
From Jordan Sissel's dotfiles:

Lots of command examples (especially heroku) lead command docs with '$' which
make it annoying to copy/paste, especially when there's multiple commands to copy.

This hacks around the problem by making a '$' command that simply runs
whatever arguments are passed to it. So you can copy